Avoid compile warnings about no previous definition of suite functions.
Some compilers, with certain warning flags, will warn about non-static functions which don't have a prototype declaration.

In the case of test suites in the same compilation unit as the test runner, this warning is spurious. The easiest way to avoid it is to add a prototype immediately before the function definition.