ASSERT_ENUM_EQ: Eliminate redundant evaluation, make type clearer.
Add typedef for greatest_enum_str_fun, the enum-to-str function expected
by ASSERT_ENUM_EQ. Add an explicit conversion, so that using
ASSERT_ENUM_EQ incorrectly will give a better type error.

Note that in C++, the stricter typechecking for enums means that thet
enum_str function's argument should be int, and it will need to be cast
back to the enum value. (In C, having the greatest_enum_str_fun argument
be the actual enum type would not produce a warning.)