Address warning: "format string is not a string literal".
Rather than saving the format string to a `const char *` and then
passing it to fprintf, just expand FMT in place twice. Since the format
string should just be a string literal to concatenate to the format
string in `GREATEST_ASSERT_EQ_FMTm`, multiple evaluation should not be a
problem.

Also, note in the comment that it will be expanded multiple times
on assertion failure.

Fixes #55.