Add `greatest_set_test_suffix(const char *name)`.
This can be used to set a suffix when printing the name of the next
test. It is cleared afterward, whether or not the test was actually run,
so the caller should be able to stack-allocate a buffer for the suffix.

Note that this is not currently searched by `test_filter`/`-t` or
`test_exclude`/`-x` -- see issue #73.

Closes #63.