Use `abort()` rather than `raise(SIGABRT)`. Rename to 'abort_on_fail'.
The C89 standard requires that `abort()` returns to the "host
environment", which is meaningful even in an embedded context without
any OS. The specifics of signal handling (in particular, which signals
exist) are less portable.

`abort()` works fine for breaking on failures in a debugger.

To test: run `./example` with `-a` in gdb.