Add `greatest_list_only()`. Use `greatest_set_flag` internally.
The stable public interface should be:

  - greatest_stop_at_first_fail()
  - greatest_abort_on_fail()
  - greatest_list_only()

with `greatest_set_flag` (and the flag names) as an implementation
detail, but that was added to the public API already. There wasn't
a way to set the list flag without using command-line arguments.
Since it prints suites/tests in a sufficiently regular format that a
non-CLI host program could scrape them, it should probably be an
option.