Implementation of name + suffix buffering (for substring match).
Instead of printing the test name as "%s%s%s", with the test name
and either `["", ""]` or `["_", suffix]`, allocate a buffer in
`greatest_info` for appending an optional suffix to the test name,
and search that buffer when doing test name filtering & exclusion.

This has the unexpected benefit of only stringifying the test name once,
in `RUN_TEST`/`RUN_TEST1`/`RUN_TESTp`, rather than multiple
times. (Though that could have been saved in a variable in the expansion
of those macros and referenced instead.)

I'm not sure what a good default size for the test name + suffix
buffer size is, but it can be configured before `#include`ing
`greatest.h` in cases where memory is tight.

The obvious implementation would use `snprintf`, but that isn't
available with `-std=c89`. Since only '_' and a suffix string
need to be appended (with truncation), just implement it using
`strncat`.

Add an example program, `example_trunc.c`, that shows that the
name buffer is truncated safely when the buffer is very small.
(Also print sizeof(greatest_info).)

Closes #73.