Make the `greatest_run_info` global struct static. Interface change.
This might be a good intermediate step toward making the
`greatest_run_info` struct static, if that's worth doing.

This is an interface change, so it will have to wait to 2.0.0,
and it's unclear if it's worth merging -- it may make other
features harder to implement, for little gain. It was an
interesting exercise, though.

It also frees up some space, since it replaces `greatest_info.`
with `g->`, so lines that reference several field in a row could
be packed together.

This depends on using a "tentative" forward declaration of
the static `greatest_run_info` struct.