Correct linear congruential RNG parameters, to fix shuffling.
Previously, certain rare combinations of test counts and input states
could prevent the linear congruential RNG from having its full period
before repeating values, which could cause tests to be run multiple
times or skipped during shuffling.

This changes the prime used (prng->c) to `(2 ** 31) - 1`, which is the
largest prime that fits in a signed 32-bit int. It also changes the
multiplier (prng->a) so prng->a - 1 is always a muliple of 4, as well as
relatively prime to prng->c, even when the input state is 0. Because the
seed is masked to its lower 29 bits, the largest seed possible produces
a multiplier of 2147483645, which is less than 2147483647, and thus
relatively prime (since 2147483647 is prime).

For an explanation of why this matters, see Knuth 3.2.1.2 Theorem A, and
note that in this case m is the smallest power of 2 greater than or
equal to the test count (since it's discarding any between the test
count and the m and redrawing).

Fixes #90.